# docs/guide/public/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Grantline API",
    "version": "0.1.0-dev",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "description": "Single-organization, cookie-authenticated API. HTTPS sessions use Secure, HttpOnly, SameSite=Lax cookies. No personal API tokens in v0.1."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "session": []
    }
  ],
  "paths": {
    "/status": {
      "get": {
        "summary": "Setup state and sign-in capabilities",
        "operationId": "get_status",
        "tags": [
          "status"
        ],
        "description": "Available before sign-in.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/auth/setup": {
      "post": {
        "summary": "Create organization and first Owner once",
        "operationId": "post_auth_setup",
        "tags": [
          "auth"
        ],
        "description": "Available before sign-in. Requires configured public Origin. Signed-in writes also require X-Grantline-CSRF from GET /auth/session.",
        "responses": {
          "201": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Setup"
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Start password session; privileged accounts require TOTP",
        "operationId": "post_auth_login",
        "tags": [
          "auth"
        ],
        "description": "Available before sign-in. Requires configured public Origin. Signed-in writes also require X-Grantline-CSRF from GET /auth/session.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        }
      }
    },
    "/auth/accept": {
      "post": {
        "summary": "Consume single-use invitation/recovery link",
        "operationId": "post_auth_accept",
        "tags": [
          "auth"
        ],
        "description": "Available before sign-in. Requires configured public Origin. Signed-in writes also require X-Grantline-CSRF from GET /auth/session.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptTicket"
              }
            }
          }
        }
      }
    },
    "/auth/forgot": {
      "post": {
        "summary": "Queue recovery mail without exposing account membership",
        "operationId": "post_auth_forgot",
        "tags": [
          "auth"
        ],
        "description": "Available before sign-in. Requires configured public Origin. Signed-in writes also require X-Grantline-CSRF from GET /auth/session.",
        "responses": {
          "202": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Email": {
                    "type": "string"
                  }
                },
                "required": [
                  "Email"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/auth/session": {
      "get": {
        "summary": "Current account and CSRF token",
        "operationId": "get_auth_session",
        "tags": [
          "auth"
        ],
        "description": "Minimum role: viewer.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/mfa": {
      "get": {
        "summary": "TOTP enrollment state and initial setup key",
        "operationId": "get_auth_mfa",
        "tags": [
          "auth"
        ],
        "description": "Minimum role: viewer.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Verify TOTP and rotate pending session",
        "operationId": "post_auth_mfa",
        "tags": [
          "auth"
        ],
        "description": "Minimum role: viewer. Requires configured public Origin. Signed-in writes also require X-Grantline-CSRF from GET /auth/session.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Grantline-CSRF",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Code": {
                    "type": "string"
                  }
                },
                "required": [
                  "Code"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/auth/logout": {
      "post": {
        "summary": "Revoke current session",
        "operationId": "post_auth_logout",
        "tags": [
          "auth"
        ],
        "description": "Minimum role: viewer. Requires configured public Origin. Signed-in writes also require X-Grantline-CSRF from GET /auth/session.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Grantline-CSRF",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/auth/password": {
      "post": {
        "summary": "Change password and revoke all sessions",
        "operationId": "post_auth_password",
        "tags": [
          "auth"
        ],
        "description": "Minimum role: viewer. Requires configured public Origin. Signed-in writes also require X-Grantline-CSRF from GET /auth/session.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Grantline-CSRF",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Current": {
                    "type": "string"
                  },
                  "Password": {
                    "type": "string"
                  }
                },
                "required": [
                  "Current",
                  "Password"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "summary": "List members",
        "operationId": "get_users",
        "tags": [
          "users"
        ],
        "description": "Minimum role: viewer.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "patch": {
        "summary": "Change role or disable; revokes sessions",
        "operationId": "patch_users_id",
        "tags": [
          "users"
        ],
        "description": "Minimum role: admin. Requires configured public Origin. Signed-in writes also require X-Grantline-CSRF from GET /auth/session.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Grantline-CSRF",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Role": {
                    "type": "string"
                  },
                  "Disabled": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "Role",
                  "Disabled"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/invitations": {
      "post": {
        "summary": "Create single-use account link, optionally email",
        "operationId": "post_invitations",
        "tags": [
          "invitations"
        ],
        "description": "Minimum role: admin. Requires configured public Origin. Signed-in writes also require X-Grantline-CSRF from GET /auth/session.",
        "responses": {
          "201": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Grantline-CSRF",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ticket"
              }
            }
          }
        }
      }
    },
    "/recovery": {
      "post": {
        "summary": "Create single-use account link, optionally email",
        "operationId": "post_recovery",
        "tags": [
          "recovery"
        ],
        "description": "Minimum role: admin. Requires configured public Origin. Signed-in writes also require X-Grantline-CSRF from GET /auth/session.",
        "responses": {
          "201": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Grantline-CSRF",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ticket"
              }
            }
          }
        }
      }
    },
...
    "/reports/import": {
      "post": {
        "summary": "Import historical report without creating connections",
        "operationId": "post_reports_import",
        "tags": [
          "reports"
        ],
        "description": "Minimum role: admin. Requires configured public Origin. Signed-in writes also require X-Grantline-CSRF from GET /auth/session.",
        "responses": {
          "201": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Grantline-CSRF",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "/docs/reference/report.schema.json"
              }
            }
          }
        }
      }
    },
...
    "/integrations": {
      "get": {
        "summary": "List connections without credentials",
        "operationId": "get_integrations",
        "tags": [
          "integrations"
        ],
        "description": "Minimum role: viewer.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create encrypted connection",
        "operationId": "post_integrations",
        "tags": [
          "integrations"
        ],
        "description": "Minimum role: admin. Requires configured public Origin. Signed-in writes also require X-Grantline-CSRF from GET /auth/session.",
        "responses": {
          "201": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Grantline-CSRF",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Integration"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "grantline_session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "Setup": {
        "type": "object",
        "properties": {
          "Token": {
            "type": "string"
          },
          "Organization": {
            "type": "string"
          },
          "Email": {
            "type": "string"
          },
          "Name": {
            "type": "string"
          },
          "Password": {
            "type": "string"
          }
        },
        "required": [
          "Token",
          "Organization",
          "Email",
          "Name",
          "Password"
        ],
        "additionalProperties": false
      },
      "Login": {
        "type": "object",
        "properties": {
          "Email": {
            "type": "string"
          },
          "Password": {
            "type": "string"
          }
        },
        "required": [
          "Email",
          "Password"
        ],
        "additionalProperties": false
      },
      "AcceptTicket": {
        "type": "object",
        "properties": {
          "Token": {
            "type": "string"
          },
          "Name": {
            "type": "string"
          },
          "Password": {
            "type": "string"
          }
        },
        "required": [
          "Token",
          "Name",
          "Password"
        ],
        "additionalProperties": false
      },
...
    }
  }
}